Feature code cleanup: type safety, immutability, decomposition, tests
Systemic:
- typedOnEvent<TRegistry>() utility eliminates payload-as casts in all 15 stores
- Typed sendCommand<K>() with Pick<Registry, ...> in all renderers
- Immutable Map updates (spread before .set()) in folders, pinned-tabs, tab-customization

Per-feature fixes:
- sidebar: decompose 1779-line god component into 8 focused files (~310 lines remaining)
- sidebar: memoize pinnedTabIds, create SidebarDragContext (not yet wired)
- command-palette: replace visible/closing booleans with status enum
- workspaces: TABS_SET_WORKSPACE command replaces direct Tab mutation
- workspaces: fix FadePresence double-rAF leak
- local-web-app: override pattern replaces useEffect state sync anti-pattern
- local-web-app: add missing awaits on async startProcess/killProcess
- tab-customization: override pattern replaces useEffect state sync
- domain-css: fix abbreviated params, wrap fs.watch async in .catch()
- debug-server: add double-registration guard to recorder monkey-patching
- pinned-tabs: emit PINNED_TABS_CHANGED on TABS_UPDATED sync
- settings: batch fire-and-forget setSetting calls into Promise.all
- tabs: add .catch() on fetchAsDataUrl promise chain
- folders: 24 new tests covering all command handlers + immutability
