+++ design-system/tsconfig.json
@@ -28,8 +28,18 @@
     "../src/renderer/src/components/BuiltInPage.tsx",
     "../src/renderer/src/components/SettingsLayout.tsx",
     "../src/features/**/*.renderer.tsx",
+    "../src/features/sidebar/Favicon.tsx",
+    "../src/features/sidebar/TabItem.tsx",
+    "../src/features/sidebar/FolderHeader.tsx",
+    "../src/features/sidebar/FolderGroup.tsx",
+    "../src/features/sidebar/BookmarkedTree.tsx",
+    "../src/features/sidebar/PinnedTabsStrip.tsx",
+    "../src/features/sidebar/SectionDropZone.tsx",
+    "../src/features/sidebar/TabSection.tsx",
+    "../src/features/sidebar/SidebarContext.tsx",
     "../src/features/**/*.store.ts",
     "../src/features/**/*.shared.ts",
-    "../src/features/command-palette/resolve-input.ts"
+    "../src/features/command-palette/resolve-input.ts",
+    "../src/bus/types.ts"
   ]
 }
