+++ src/features/sidebar/sidebar.renderer.tsx
@@ -1,6 +1,5 @@
 import { useCallback, useEffect, useLayoutEffect, useMemo, useRef, useState } from "react";
-import { type ContextMenuItem, useContextMenu } from "../../renderer/src/components/ContextMenu";
-import { Icon } from "../../renderer/src/components/Icon";
+import { useContextMenu } from "../../renderer/src/components/ContextMenu";
 import type { FolderId, TabId, WorkspaceId } from "../../shared/types";
 import {
   APP_STATE_SET_SIDEBAR_WIDTH,
@@ -10,83 +9,49 @@ import {
 // shell-composite: read-only cross-feature store access
 import { useAppStateStore } from "../app-state/app-state.store";
 import { DownloadsSection } from "../downloads/downloads.renderer";
-import type { Folder } from "../folders/folders.shared";
-import {
-  FOLDERS_CREATE,
-  FOLDERS_REMOVE,
-  FOLDERS_RENAME,
-  FOLDERS_REORDER,
-  FOLDERS_TOGGLE_COLLAPSE,
-} from "../folders/folders.shared";
+import type { Folder, FoldersCommands } from "../folders/folders.shared";
+import { FOLDERS_CREATE } from "../folders/folders.shared";
 // shell-composite: read-only cross-feature store access
 import { useFoldersStore } from "../folders/folders.store";
 import { UpdateNotification } from "../installer/installer.renderer";
-import type { PinnedTabsCommands } from "../pinned-tabs/pinned-tabs.shared";
-import { PINNED_TABS_ACTIVATE, PINNED_TABS_TOGGLE_PIN } from "../pinned-tabs/pinned-tabs.shared";
 // shell-composite: read-only cross-feature store access
 import { usePinnedTabsStore } from "../pinned-tabs/pinned-tabs.store";
-import {
-  TAB_CUSTOMIZATION_OPEN,
-  type TabCustomizationCommands,
-} from "../tab-customization/tab-customization.shared";
-// shell-composite: read-only cross-feature store access
-import { useTabCustomizationStore } from "../tab-customization/tab-customization.store";
-import type { Tab, TabsCommands } from "../tabs/tabs.shared";
-import {
-  TABS_ACTIVATE,
-  TABS_CLEAR_EPHEMERAL,
-  TABS_CLOSE,
-  TABS_NAVIGATE,
-  TABS_REORDER,
-  TABS_TOGGLE_BOOKMARK,
-} from "../tabs/tabs.shared";
+import type { Tab } from "../tabs/tabs.shared";
+import { TABS_CLEAR_EPHEMERAL } from "../tabs/tabs.shared";
 // shell-composite: read-only cross-feature store access
 import { useTabsStore } from "../tabs/tabs.store";
 import { WorkspaceSwitcher } from "../workspaces/workspaces.renderer";
 import type { Workspace } from "../workspaces/workspaces.shared";
 // shell-composite: read-only cross-feature store access
 import { useWorkspacesStore } from "../workspaces/workspaces.store";
+import { PinnedTabsStrip } from "./PinnedTabsStrip";
+import { TabSection } from "./TabSection";
 import { useSidebarStore } from "./sidebar.store";
+
+// ── Re-exports for external consumers ───────────────────────────
+
+export { hashToHue, Favicon } from "./Favicon";
+export { TabItem } from "./TabItem";
+export { PinnedTabsStrip } from "./PinnedTabsStrip";
+
+// ── Shared types ────────────────────────────────────────────────
+
+export type TreeItem =
+  | { type: "tab"; tab: Tab }
+  | { type: "folder"; folder: Folder; children: TreeItem[] };
 
 // ── Typed sendCommand ───────────────────────────────────────────
 
-type SidebarUsedCommands = Pick<
-  TabsCommands,
-  | typeof TABS_ACTIVATE
-  | typeof TABS_CLOSE
-  | typeof TABS_NAVIGATE
-  | typeof TABS_CLEAR_EPHEMERAL
-  | typeof TABS_REORDER
-> &
-  Pick<PinnedTabsCommands, typeof PINNED_TABS_ACTIVATE> &
-  Pick<TabCustomizationCommands, typeof TAB_CUSTOMIZATION_OPEN>;
+type SidebarUsedCommands = Pick<FoldersCommands, typeof FOLDERS_CREATE>;
 
-function sendCommand<K extends keyof SidebarUsedCommands>(
+function sendFolderCommand<K extends keyof SidebarUsedCommands>(
   name: K,
   payload: SidebarUsedCommands[K]["payload"],
 ) {
   window.chiaroscuro.sendCommand(name, payload);
 }
 
-function sendFolderCommand(name: string, payload: unknown) {
-  window.chiaroscuro.sendCommand(name, payload);
-}
-
-// ── Helpers ─────────────────────────────────────────────────────
-
-export function hashToHue(str: string): number {
-  let hash = 0;
-  for (let i = 0; i < str.length; i++) {
-    hash = ((hash << 5) - hash + str.charCodeAt(i)) | 0;
-  }
-  return Math.abs(hash) % 360;
-}
-
 // ── Tree building ───────────────────────────────────────────────
-
-type TreeItem =
-  | { type: "tab"; tab: Tab }
-  | { type: "folder"; folder: Folder; children: TreeItem[] };
 
 function buildBookmarkedTree(
   tabs: Tab[],
@@ -128,26 +93,6 @@ function buildBookmarkedTree(
   }
 
   return buildLevel(null);
-}
-
-function findFirstTabId(items: TreeItem[]): TabId | null {
-  for (const item of items) {
-    if (item.type === "tab") return item.tab.id;
-    const found = findFirstTabId(item.children);
-    if (found) return found;
-  }
-  return null;
-}
-
-function findLastTabId(items: TreeItem[]): TabId | null {
-  for (let i = items.length - 1; i >= 0; i--) {
-    const item = items[i];
-    if (!item) continue;
-    if (item.type === "tab") return item.tab.id;
-    const found = findLastTabId(item.children);
-    if (found) return found;
-  }
-  return null;
 }
 
 // ── Hooks ───────────────────────────────────────────────────────
@@ -203,1281 +148,7 @@ function useWorkspaceSlide(activeWorkspaceId: WorkspaceId | null, workspaces: Wo
   return transition;
 }
 
-// ── Built-in page icons ─────────────────────────────────────────
-
-import type { FaSolidIcon } from "../../shared/fa-icons.generated";
-
-const builtInIcons: Record<string, FaSolidIcon> = {
-  "app:settings": "gear",
-};
-
-// ── Components ──────────────────────────────────────────────────
-
-function LetterAvatar({ label, hue }: { label: string; hue: number }) {
-  return (
-    <div
-      className="shrink-0 flex items-center justify-center rounded-full"
-      style={{
-        width: 16,
-        height: 16,
-        fontSize: "var(--text-xs)",
-        fontWeight: 600,
-        fontFamily: "var(--font-sans)",
-        color: "var(--glass-text-primary)",
-        background: `oklch(0.55 0.15 ${hue})`,
-      }}
-    >
-      {label}
-    </div>
-  );
-}
-
-export function Favicon({ tab }: { tab: Pick<Tab, "favicon" | "title" | "url"> }) {
-  const [imgFailed, setImgFailed] = useState(false);
-  const [retries, setRetries] = useState(0);
-  const [lastFavicon, setLastFavicon] = useState(tab.favicon);
-
-  // Reset failure state when favicon URL changes
-  if (tab.favicon !== lastFavicon) {
-    setLastFavicon(tab.favicon);
-    setImgFailed(false);
-    setRetries(0);
-  }
-
-  // Retry localhost favicons (dev servers may not be ready yet on restore)
-  useEffect(() => {
-    if (!imgFailed || retries >= 3 || !tab.favicon) return;
-    try {
-      const u = new URL(tab.favicon);
-      if (u.hostname !== "localhost" && u.hostname !== "127.0.0.1") return;
-    } catch {
-      return;
-    }
-    const timer = setTimeout(() => {
-      setImgFailed(false);
-      setRetries((r) => r + 1);
-    }, 2000);
-    return () => clearTimeout(timer);
-  }, [imgFailed, retries, tab.favicon]);
-
-  const letter = tab.title?.[0]?.toUpperCase() || tab.url?.[0]?.toUpperCase() || "?";
-  const hue = hashToHue(tab.url || tab.title);
-
-  if (tab.favicon && !imgFailed) {
-    return (
-      <img
-        src={tab.favicon}
-        alt=""
-        className="shrink-0 rounded-full"
-        style={{ width: 16, height: 16 }}
-        onError={() => setImgFailed(true)}
-      />
-    );
-  }
-
-  const builtInIcon = builtInIcons[tab.url];
-  if (builtInIcon) {
-    return (
-      <div className="shrink-0 flex items-center justify-center" style={{ width: 16, height: 16 }}>
-        <Icon name={builtInIcon} css={{ fontSize: 12, color: "var(--glass-text-muted)" }} />
-      </div>
-    );
-  }
-
-  return <LetterAvatar label={letter} hue={hue} />;
-}
-
-export function TabItem({
-  tab,
-  isActive,
-  isEphemeral,
-  isPinned,
-  exiting,
-  isBookmarkedSection,
-  folderId,
-  dragTabIdRef,
-  dragFolderIdRef,
-  isDragged,
-  isDragging,
-  onBeforeReorder,
-  lastSwapRef,
-  lastSwapTimeRef,
-  lastFolderSwapRef,
-  lastFolderSwapTimeRef,
-  disableEntryAnimation,
-  onContextMenu,
-}: {
-  tab: Tab;
-  isActive: boolean;
-  isEphemeral: boolean;
-  isPinned?: boolean;
-  exiting?: boolean;
-  isBookmarkedSection: boolean;
-  folderId?: FolderId | null;
-  dragTabIdRef: React.RefObject<TabId | null>;
-  dragFolderIdRef?: React.RefObject<FolderId | null>;
-  isDragged: boolean;
-  isDragging: boolean;
-  onBeforeReorder: () => void;
-  lastSwapRef: React.RefObject<{ targetId: TabId; position: string } | null>;
-  lastSwapTimeRef: React.RefObject<number>;
-  lastFolderSwapRef?: React.RefObject<{ targetId: string; position: string } | null>;
-  lastFolderSwapTimeRef?: React.RefObject<number>;
-  disableEntryAnimation?: boolean;
-  onContextMenu?: (items: ContextMenuItem[], e: React.MouseEvent) => void;
-}) {
-  const customization = useTabCustomizationStore((s) => s.customizations.get(tab.id));
-  const customTitle = customization?.title;
-  const mountedRef = useRef(false);
-  const elRef = useRef<HTMLDivElement>(null);
-  useEffect(() => {
-    const t = setTimeout(() => {
-      mountedRef.current = true;
-    }, 200);
-    return () => clearTimeout(t);
-  }, []);
-
-  const handleClick = () => {
-    sendCommand(TABS_ACTIVATE, { tabId: tab.id });
-  };
-
-  const handleClose = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    sendCommand(TABS_CLOSE, { tabId: tab.id });
-  };
-
-  const handleContextMenu = (e: React.MouseEvent) => {
-    if (!onContextMenu) return;
-    const items: ContextMenuItem[] = [];
-    if (isEphemeral) {
-      items.push({
-        label: "Bookmark",
-        icon: "bookmark",
-        onSelect: () => sendFolderCommand(TABS_TOGGLE_BOOKMARK, { tabId: tab.id }),
-      });
-    }
-    if (isBookmarkedSection && !isPinned) {
-      items.push({
-        label: "Remove bookmark",
-        icon: "bookmark",
-        onSelect: () => sendFolderCommand(TABS_TOGGLE_BOOKMARK, { tabId: tab.id }),
-      });
-    }
-    if (!isPinned) {
-      items.push({
-        label: "Pin tab",
-        icon: "thumbtack",
-        onSelect: () => sendFolderCommand(PINNED_TABS_TOGGLE_PIN, { tabId: tab.id }),
-      });
-    }
-    if (isPinned) {
-      items.push({
-        label: "Unpin tab",
-        icon: "thumbtack-slash",
-        onSelect: () => sendFolderCommand(PINNED_TABS_TOGGLE_PIN, { tabId: tab.id }),
-      });
-      items.push({
-        label: "Close tab",
-        icon: "xmark",
-        onSelect: () => sendCommand(TABS_CLOSE, { tabId: tab.id }),
-      });
-    }
-    if (!isEphemeral) {
-      items.push({
-        label: "Customize tab",
-        icon: "sliders",
-        onSelect: () => sendCommand(TAB_CUSTOMIZATION_OPEN, { tabId: tab.id }),
-      });
-    }
-    const fixedUrl = tab.fixedUrl;
-    if (fixedUrl && tab.url !== fixedUrl) {
-      items.push({
-        label: "Restore original URL",
-        icon: "arrow-rotate-left",
-        onSelect: () => sendCommand(TABS_NAVIGATE, { tabId: tab.id, url: fixedUrl }),
-      });
-    }
-    onContextMenu(items, e);
-  };
-
-  const handleDragStart = (e: React.DragEvent) => {
-    e.dataTransfer.effectAllowed = "move";
-    e.dataTransfer.setData("text/plain", tab.id);
-    dragTabIdRef.current = tab.id;
-    // Hide browser's default drag ghost
-    const img = new Image();
-    img.src = "data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7";
-    e.dataTransfer.setDragImage(img, 0, 0);
-    lastSwapRef.current = null;
-    lastSwapTimeRef.current = 0;
-  };
-
-  const handleDragOver = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.dataTransfer.dropEffect = "move";
-
-    // Handle folder being dragged over this tab
-    const draggingFolderId = dragFolderIdRef?.current;
-    if (draggingFolderId && isBookmarkedSection) {
-      if (
-        !lastFolderSwapRef ||
-        !lastFolderSwapTimeRef ||
-        Date.now() - lastFolderSwapTimeRef.current < 100
-      )
-        return;
-      const rect = elRef.current?.getBoundingClientRect();
-      if (!rect) return;
-      const position = e.clientY > rect.top + rect.height / 2 ? "after" : "before";
-      if (
-        lastFolderSwapRef.current?.targetId === tab.id &&
-        lastFolderSwapRef.current?.position === position
-      )
-        return;
-      lastFolderSwapRef.current = { targetId: tab.id, position };
-      lastFolderSwapTimeRef.current = Date.now();
-      onBeforeReorder();
-      sendFolderCommand(FOLDERS_REORDER, {
-        folderId: draggingFolderId,
-        parentFolderId: folderId ?? null,
-        targetTabId: tab.id,
-        position,
-      });
-      return;
-    }
-
-    // Handle tab being dragged over this tab
-    const tabId = dragTabIdRef.current;
-    if (!tabId || tabId === tab.id) return;
-    if (Date.now() - lastSwapTimeRef.current < 100) return;
-    const rect = elRef.current?.getBoundingClientRect();
-    if (!rect) return;
-    const position = e.clientY > rect.top + rect.height / 2 ? "after" : "before";
-    if (lastSwapRef.current?.targetId === tab.id && lastSwapRef.current?.position === position)
-      return;
-    lastSwapRef.current = { targetId: tab.id, position };
-    lastSwapTimeRef.current = Date.now();
-    onBeforeReorder();
-    sendCommand(TABS_REORDER, {
-      tabId,
-      targetBookmarked: isBookmarkedSection,
-      targetTabId: tab.id,
-      position,
-      targetFolderId: folderId ?? null,
-    });
-  };
-
-  const handleDrop = (e: React.DragEvent) => {
-    e.preventDefault();
-  };
-
-  return (
-    // biome-ignore lint/a11y/useKeyWithClickEvents: chrome elements are not keyboard-navigable
-    <div
-      ref={elRef}
-      data-tab-id={tab.id}
-      data-ephemeral={isEphemeral ? "" : undefined}
-      draggable={!exiting}
-      className={`${isDragging ? "" : "group"} relative flex items-center cursor-pointer transition-colors duration-150 ${isDragging ? "" : "hover:bg-glass-hover hover:text-glass-text-hover active:bg-glass-pressed active:text-glass-text-pressed"}`}
-      style={{
-        gap: "0.625rem",
-        padding: "0.375rem 0.75rem",
-        margin: "0.25rem 0.375rem",
-        borderRadius: "var(--radius-md)",
-        background: isDragged
-          ? "oklch(var(--accent-L) var(--accent-C) var(--accent-hue, 250) / 0.06)"
-          : isActive
-            ? "var(--glass-active)"
-            : undefined,
-        boxShadow: isDragged
-          ? "inset 0 0 0 1px oklch(var(--accent-L) var(--accent-C) var(--accent-hue, 250) / 0.25)"
-          : isActive
-            ? "var(--shadow-subtle)"
-            : undefined,
-        zIndex: isDragged ? 10 : undefined,
-        pointerEvents: exiting ? "none" : undefined,
-        animation: exiting
-          ? "tab-out 200ms cubic-bezier(0.4, 0, 1, 1) forwards"
-          : mountedRef.current || disableEntryAnimation
-            ? undefined
-            : "tab-in 200ms cubic-bezier(0, 0, 0.2, 1) both",
-      }}
-      onClick={handleClick}
-      onContextMenu={handleContextMenu}
-      onDragStart={handleDragStart}
-      onDragOver={handleDragOver}
-      onDrop={handleDrop}
-    >
-      <div className="shrink-0">
-        <Favicon tab={tab} />
-      </div>
-      <span
-        className={`flex-1 min-w-0 truncate group-hover:text-glass-text-hover group-hover:mr-5 group-active:text-glass-text-pressed ${isActive ? "text-glass-text-primary" : isEphemeral ? "text-glass-text-muted" : "text-glass-text-default"}`}
-        style={{
-          fontSize: "var(--text-base)",
-          fontWeight: isActive ? 500 : undefined,
-          transition: "margin var(--duration-fast) var(--ease-in-out)",
-        }}
-      >
-        {customTitle || tab.title || tab.url}
-      </span>
-      <button
-        type="button"
-        className="absolute flex opacity-0 group-hover:opacity-100 items-center justify-center bg-transparent text-glass-text-hint transition-[opacity,color] duration-150 hover:text-destructive"
-        style={{
-          right: "0.375rem",
-          top: "50%",
-          transform: "translateY(-50%)",
-          width: 24,
-          height: 24,
-          borderRadius: "var(--radius-sm)",
-          cursor: "pointer",
-          border: "none",
-        }}
-        tabIndex={-1}
-        onClick={handleClose}
-        aria-label="Close tab"
-        data-tip="Close tab"
-      >
-        <Icon name="xmark" css={{ fontSize: 10 }} />
-      </button>
-    </div>
-  );
-}
-
-// ── Folder Components ───────────────────────────────────────────
-
-function FolderHeader({
-  folder,
-  isRenaming,
-  isDragging,
-  dragTabIdRef,
-  dragFolderIdRef,
-  depth,
-  onBeforeReorder,
-  lastFolderSwapRef,
-  lastFolderSwapTimeRef,
-  lastSwapRef,
-  lastSwapTimeRef,
-  firstSubtreeTabId,
-  lastSubtreeTabId,
-  onContextMenu,
-}: {
-  folder: Folder;
-  isRenaming: boolean;
-  isDragging: boolean;
-  dragTabIdRef: React.RefObject<TabId | null>;
-  dragFolderIdRef: React.RefObject<FolderId | null>;
-  depth: number;
-  onBeforeReorder: () => void;
-  lastFolderSwapRef: React.RefObject<{ targetId: FolderId; position: string } | null>;
-  lastFolderSwapTimeRef: React.RefObject<number>;
-  lastSwapRef: React.RefObject<{ targetId: TabId; position: string } | null>;
-  lastSwapTimeRef: React.RefObject<number>;
-  firstSubtreeTabId: TabId | null;
-  lastSubtreeTabId: TabId | null;
-  onContextMenu?: (items: ContextMenuItem[], e: React.MouseEvent) => void;
-}) {
-  const inputRef = useRef<HTMLInputElement>(null);
-  const [renameValue, setRenameValue] = useState(folder.name);
-  const [dropOver, setDropOver] = useState(false);
-
-  // Reset drop highlight when global drag ends (fallback for lost events)
-  useEffect(() => {
-    if (!isDragging) setDropOver(false);
-  }, [isDragging]);
-
-  useEffect(() => {
-    if (isRenaming && inputRef.current) {
-      inputRef.current.focus();
-      inputRef.current.select();
-      setRenameValue(folder.name);
-    }
-  }, [isRenaming, folder.name]);
-
-  const commitRename = () => {
-    const trimmed = renameValue.trim();
-    if (trimmed && trimmed !== folder.name) {
-      sendFolderCommand(FOLDERS_RENAME, { folderId: folder.id, name: trimmed });
-    }
-    useFoldersStore.setState({ renamingFolderId: null });
-  };
-
-  const handleHeaderClick = () => {
-    if (!isRenaming) {
-      sendFolderCommand(FOLDERS_TOGGLE_COLLAPSE, { folderId: folder.id });
-    }
-  };
-
-  const handleDoubleClick = () => {
-    useFoldersStore.setState({ renamingFolderId: folder.id });
-  };
-
-  const handleRemove = (e: React.MouseEvent) => {
-    e.stopPropagation();
-    sendFolderCommand(FOLDERS_REMOVE, { folderId: folder.id });
-  };
-
-  const handleFolderContextMenu = (e: React.MouseEvent) => {
-    if (!onContextMenu) return;
-    e.stopPropagation();
-    onContextMenu(
-      [
-        {
-          label: "Add subfolder",
-          icon: "folder-plus",
-          onSelect: () => sendFolderCommand(FOLDERS_CREATE, { parentFolderId: folder.id }),
-        },
-      ],
-      e,
-    );
-  };
-
-  const handleFolderDragStart = (e: React.DragEvent) => {
-    if (isRenaming) {
-      e.preventDefault();
-      return;
-    }
-    e.dataTransfer.effectAllowed = "move";
-    e.dataTransfer.setData("text/x-folder", folder.id);
-    dragFolderIdRef.current = folder.id;
-    // Hide browser's default drag ghost
-    const img = new Image();
-    img.src = "data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7";
-    e.dataTransfer.setDragImage(img, 0, 0);
-    lastFolderSwapRef.current = null;
-    lastFolderSwapTimeRef.current = 0;
-  };
-
-  const elRef = useRef<HTMLDivElement>(null);
-
-  const handleDragOver = (e: React.DragEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    e.dataTransfer.dropEffect = "move";
-
-    const draggingFolderId = dragFolderIdRef.current;
-    if (draggingFolderId) {
-      // Folder-on-folder: reorder as siblings or nest
-      if (draggingFolderId === folder.id) return;
-      if (Date.now() - lastFolderSwapTimeRef.current < 100) return;
-      const rect = elRef.current?.getBoundingClientRect();
-      if (!rect) return;
-      const relY = (e.clientY - rect.top) / rect.height;
-      if (relY > 0.25 && relY < 0.75) {
-        // Middle zone: nest into this folder
-        setDropOver(true);
-        if (
-          lastFolderSwapRef.current?.targetId === folder.id &&
-          lastFolderSwapRef.current?.position === "inside"
-        )
-          return;
-        lastFolderSwapRef.current = { targetId: folder.id, position: "inside" };
-        lastFolderSwapTimeRef.current = Date.now();
-        sendFolderCommand(FOLDERS_REORDER, {
-          folderId: draggingFolderId,
-          parentFolderId: folder.id,
-        });
-      } else {
-        // Top/bottom zone: reorder as sibling
-        setDropOver(false);
-        const position = relY >= 0.75 ? "after" : "before";
-        if (
-          lastFolderSwapRef.current?.targetId === folder.id &&
-          lastFolderSwapRef.current?.position === position
-        )
-          return;
-        lastFolderSwapRef.current = { targetId: folder.id, position };
-        lastFolderSwapTimeRef.current = Date.now();
-        sendFolderCommand(FOLDERS_REORDER, {
-          folderId: draggingFolderId,
-          targetFolderId: folder.id,
-          position,
-          parentFolderId: folder.parentFolderId,
-        });
-      }
-    } else {
-      // Tab being dragged over folder header
-      const tabId = dragTabIdRef.current;
-      if (!tabId) return;
-      const rect = elRef.current?.getBoundingClientRect();
-      if (!rect) return;
-      const relY = (e.clientY - rect.top) / rect.height;
-
-      if (relY > 0.25 && relY < 0.75) {
-        // Middle zone: show drop highlight (nesting happens on drop)
-        setDropOver(true);
-      } else {
-        // Top/bottom zone: live reorder to position adjacent to folder
-        setDropOver(false);
-        if (Date.now() - lastSwapTimeRef.current < 100) return;
-
-        if (relY <= 0.25 && firstSubtreeTabId) {
-          // Before folder
-          if (
-            lastSwapRef.current?.targetId === firstSubtreeTabId &&
-            lastSwapRef.current?.position === "before"
-          )
-            return;
-          lastSwapRef.current = { targetId: firstSubtreeTabId, position: "before" };
-          lastSwapTimeRef.current = Date.now();
-          onBeforeReorder();
-          sendCommand(TABS_REORDER, {
-            tabId,
-            targetBookmarked: true,
-            targetTabId: firstSubtreeTabId,
-            position: "before",
-            targetFolderId: folder.parentFolderId ?? null,
-          });
-        } else if (relY >= 0.75 && lastSubtreeTabId) {
-          // After folder
-          if (
-            lastSwapRef.current?.targetId === lastSubtreeTabId &&
-            lastSwapRef.current?.position === "after"
-          )
-            return;
-          lastSwapRef.current = { targetId: lastSubtreeTabId, position: "after" };
-          lastSwapTimeRef.current = Date.now();
-          onBeforeReorder();
-          sendCommand(TABS_REORDER, {
-            tabId,
-            targetBookmarked: true,
-            targetTabId: lastSubtreeTabId,
-            position: "after",
-            targetFolderId: folder.parentFolderId ?? null,
-          });
-        } else {
-          // No subtree tabs — treat as nest zone
-          setDropOver(true);
-        }
-      }
-    }
-  };
-
-  const handleDragLeave = () => {
-    setDropOver(false);
-  };
-
-  const handleDrop = (e: React.DragEvent) => {
-    e.preventDefault();
-    setDropOver(false);
-    // Tab drop — only nest if in middle zone
-    const tabId = dragTabIdRef.current;
-    if (tabId) {
-      const rect = elRef.current?.getBoundingClientRect();
-      if (rect) {
-        const relY = (e.clientY - rect.top) / rect.height;
-        if (relY > 0.25 && relY < 0.75) {
-          onBeforeReorder();
-          sendCommand(TABS_REORDER, {
-            tabId,
-            targetBookmarked: true,
-            targetFolderId: folder.id,
-          });
-        }
-        // Top/bottom zones already handled by dragOver live reorder
-      }
-    }
-    // Folder drop handled in dragOver already (live reorder)
-  };
-
-  const isDraggedFolder = isDragging && dragFolderIdRef.current === folder.id;
-
-  return (
-    // biome-ignore lint/a11y/useKeyWithClickEvents: chrome elements are not keyboard-navigable
-    <div
-      ref={elRef}
-      draggable={!isRenaming}
-      className={`${isDragging ? "" : "group"} relative flex items-center cursor-pointer transition-colors duration-150 ${isDragging ? "" : "hover:bg-glass-hover hover:text-glass-text-hover active:bg-glass-pressed active:text-glass-text-pressed"}`}
-      style={{
-        gap: "0.625rem",
-        padding: "0.375rem 0.75rem",
-        paddingLeft: "0.75rem",
-        margin: "0.25rem 0.375rem",
-        borderRadius: "var(--radius-md)",
-        background: dropOver
-          ? "oklch(var(--accent-L) var(--accent-C) var(--accent-hue, 250) / 0.1)"
-          : isDraggedFolder
-            ? "oklch(var(--accent-L) var(--accent-C) var(--accent-hue, 250) / 0.06)"
-            : undefined,
-        boxShadow: dropOver
-          ? "inset 0 0 0 1px oklch(var(--accent-L) var(--accent-C) var(--accent-hue, 250) / 0.3)"
-          : isDraggedFolder
-            ? "inset 0 0 0 1px oklch(var(--accent-L) var(--accent-C) var(--accent-hue, 250) / 0.25)"
-            : undefined,
-        zIndex: isDraggedFolder ? 10 : undefined,
-      }}
-      onClick={handleHeaderClick}
-      onContextMenu={handleFolderContextMenu}
-      onDoubleClick={handleDoubleClick}
-      onDragStart={handleFolderDragStart}
-      onDragOver={handleDragOver}
-      onDragLeave={handleDragLeave}
-      onDrop={handleDrop}
-    >
-      <div
-        className="text-glass-text-default group-hover:text-glass-text-hover group-active:text-glass-text-pressed"
-        style={{ position: "relative", width: 16, height: 16, transition: "color 150ms" }}
-      >
-        <Icon
-          name="folder"
-          css={{
-            fontSize: 14,
-            position: "absolute",
-            inset: 0,
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-            opacity: folder.collapsed ? 1 : 0,
-            transform: folder.collapsed ? "scale(1)" : "scale(0.8)",
-            transition: "opacity 150ms, transform 150ms",
-          }}
-        />
-        <Icon
-          name="folder-open"
-          css={{
-            fontSize: 14,
-            position: "absolute",
-            inset: 0,
-            display: "flex",
-            alignItems: "center",
-            justifyContent: "center",
-            opacity: folder.collapsed ? 0 : 1,
-            transform: folder.collapsed ? "scale(0.8)" : "scale(1)",
-            transition: "opacity 150ms, transform 150ms",
-          }}
-        />
-      </div>
-      {isRenaming ? (
-        <input
-          ref={inputRef}
-          type="text"
-          value={renameValue}
-          onChange={(e) => setRenameValue(e.target.value)}
-          onBlur={(e) => {
-            // Ignore blur from DOM detachment during React re-renders
-            if (e.currentTarget.isConnected) commitRename();
-          }}
-          onKeyDown={(e) => {
-            if (e.key === "Enter") commitRename();
-            if (e.key === "Escape") {
-              useFoldersStore.setState({ renamingFolderId: null });
-            }
-          }}
-          className="flex-1 min-w-0 bg-transparent text-glass-text-primary outline-none"
-          style={{
-            fontSize: "var(--text-base)",
-            fontWeight: 500,
-            fontFamily: "inherit",
-            border: "none",
-            padding: 0,
-            margin: 0,
-          }}
-          onClick={(e) => e.stopPropagation()}
-        />
-      ) : (
-        <span
-          className="flex-1 min-w-0 truncate text-glass-text-default group-hover:text-glass-text-hover group-hover:mr-5 group-active:text-glass-text-pressed"
-          style={{
-            fontSize: "var(--text-base)",
-            fontWeight: 500,
-            transition: "margin var(--duration-fast) var(--ease-in-out)",
-          }}
-        >
-          {folder.name}
-        </span>
-      )}
-      {!isRenaming && (
-        <button
-          type="button"
-          className="absolute flex opacity-0 group-hover:opacity-100 items-center justify-center bg-transparent text-glass-text-hint transition-[opacity,color] duration-150 hover:text-destructive"
-          style={{
-            right: "0.375rem",
-            top: "50%",
-            transform: "translateY(-50%)",
-            width: 24,
-            height: 24,
-            borderRadius: "var(--radius-sm)",
-            cursor: "pointer",
-            border: "none",
-          }}
-          tabIndex={-1}
-          onClick={handleRemove}
-          aria-label="Remove folder"
-          data-tip="Remove folder"
-        >
-          <Icon name="xmark" css={{ fontSize: 10 }} />
-        </button>
-      )}
-    </div>
-  );
-}
-
-function FolderGroup({
-  folder,
-  items,
-  activeTabId,
-  exitingIds,
-  dragTabIdRef,
-  dragFolderIdRef,
-  isDragging,
-  onBeforeReorder,
-  lastSwapRef,
-  lastSwapTimeRef,
-  lastFolderSwapRef,
-  lastFolderSwapTimeRef,
-  disableEntryAnimation,
-  renamingFolderId,
-  depth,
-  onContextMenu,
-}: {
-  folder: Folder;
-  items: TreeItem[];
-  activeTabId: TabId | null;
-  exitingIds: Set<TabId>;
-  dragTabIdRef: React.RefObject<TabId | null>;
-  dragFolderIdRef: React.RefObject<FolderId | null>;
-  isDragging: boolean;
-  onBeforeReorder: () => void;
-  lastSwapRef: React.RefObject<{ targetId: TabId; position: string } | null>;
-  lastSwapTimeRef: React.RefObject<number>;
-  lastFolderSwapRef: React.RefObject<{ targetId: FolderId; position: string } | null>;
-  lastFolderSwapTimeRef: React.RefObject<number>;
-  disableEntryAnimation?: boolean;
-  renamingFolderId: FolderId | null;
-  depth: number;
-  onContextMenu?: (items: ContextMenuItem[], e: React.MouseEvent) => void;
-}) {
-  const firstTab = findFirstTabId(items);
-  const lastTab = findLastTabId(items);
-
-  return (
-    <div data-folder-id={folder.id}>
-      <FolderHeader
-        folder={folder}
-        isRenaming={renamingFolderId === folder.id}
-        isDragging={isDragging}
-        dragTabIdRef={dragTabIdRef}
-        dragFolderIdRef={dragFolderIdRef}
-        depth={depth}
-        onBeforeReorder={onBeforeReorder}
-        lastFolderSwapRef={lastFolderSwapRef}
-        lastFolderSwapTimeRef={lastFolderSwapTimeRef}
-        lastSwapRef={lastSwapRef}
-        lastSwapTimeRef={lastSwapTimeRef}
-        firstSubtreeTabId={firstTab}
-        lastSubtreeTabId={lastTab}
-        onContextMenu={onContextMenu}
-      />
-      {!folder.collapsed && (
-        <div style={{ paddingLeft: "0.5rem" }}>
-          {items.map((item) =>
-            item.type === "tab" ? (
-              <TabItem
-                key={item.tab.id}
-                tab={item.tab}
-                isActive={item.tab.id === activeTabId}
-                isEphemeral={false}
-                exiting={exitingIds.has(item.tab.id)}
-                isBookmarkedSection={true}
-                folderId={folder.id}
-                dragTabIdRef={dragTabIdRef}
-                dragFolderIdRef={dragFolderIdRef}
-                isDragged={item.tab.id === (isDragging ? dragTabIdRef.current : null)}
-                isDragging={isDragging}
-                onBeforeReorder={onBeforeReorder}
-                lastSwapRef={lastSwapRef}
-                lastSwapTimeRef={lastSwapTimeRef}
-                lastFolderSwapRef={lastFolderSwapRef}
-                lastFolderSwapTimeRef={lastFolderSwapTimeRef}
-                disableEntryAnimation={disableEntryAnimation}
-                onContextMenu={onContextMenu}
-              />
-            ) : (
-              <FolderGroup
-                key={item.folder.id}
-                folder={item.folder}
-                items={item.children}
-                activeTabId={activeTabId}
-                exitingIds={exitingIds}
-                dragTabIdRef={dragTabIdRef}
-                dragFolderIdRef={dragFolderIdRef}
-                isDragging={isDragging}
-                onBeforeReorder={onBeforeReorder}
-                lastSwapRef={lastSwapRef}
-                lastSwapTimeRef={lastSwapTimeRef}
-                lastFolderSwapRef={lastFolderSwapRef}
-                lastFolderSwapTimeRef={lastFolderSwapTimeRef}
-                disableEntryAnimation={disableEntryAnimation}
-                renamingFolderId={renamingFolderId}
-                depth={depth + 1}
-                onContextMenu={onContextMenu}
-              />
-            ),
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-function BookmarkedTree({
-  tree,
-  activeTabId,
-  exitingIds,
-  dragTabIdRef,
-  dragFolderIdRef,
-  isDragging,
-  onBeforeReorder,
-  lastSwapRef,
-  lastSwapTimeRef,
-  lastFolderSwapRef,
-  lastFolderSwapTimeRef,
-  disableEntryAnimation,
-  renamingFolderId,
-  onContextMenu,
-}: {
-  tree: TreeItem[];
-  activeTabId: TabId | null;
-  exitingIds: Set<TabId>;
-  dragTabIdRef: React.RefObject<TabId | null>;
-  dragFolderIdRef: React.RefObject<FolderId | null>;
-  isDragging: boolean;
-  onBeforeReorder: () => void;
-  lastSwapRef: React.RefObject<{ targetId: TabId; position: string } | null>;
-  lastSwapTimeRef: React.RefObject<number>;
-  lastFolderSwapRef: React.RefObject<{ targetId: FolderId; position: string } | null>;
-  lastFolderSwapTimeRef: React.RefObject<number>;
-  disableEntryAnimation?: boolean;
-  renamingFolderId: FolderId | null;
-  onContextMenu?: (items: ContextMenuItem[], e: React.MouseEvent) => void;
-}) {
-  return (
-    <>
-      {tree.map((item) =>
-        item.type === "tab" ? (
-          <TabItem
-            key={item.tab.id}
-            tab={item.tab}
-            isActive={item.tab.id === activeTabId}
-            isEphemeral={false}
-            exiting={exitingIds.has(item.tab.id)}
-            isBookmarkedSection={true}
-            folderId={null}
-            dragTabIdRef={dragTabIdRef}
-            dragFolderIdRef={dragFolderIdRef}
-            isDragged={item.tab.id === (isDragging ? dragTabIdRef.current : null)}
-            isDragging={isDragging}
-            onBeforeReorder={onBeforeReorder}
-            lastSwapRef={lastSwapRef}
-            lastSwapTimeRef={lastSwapTimeRef}
-            lastFolderSwapRef={lastFolderSwapRef}
-            lastFolderSwapTimeRef={lastFolderSwapTimeRef}
-            disableEntryAnimation={disableEntryAnimation}
-            onContextMenu={onContextMenu}
-          />
-        ) : (
-          <FolderGroup
-            key={item.folder.id}
-            folder={item.folder}
-            items={item.children}
-            activeTabId={activeTabId}
-            exitingIds={exitingIds}
-            dragTabIdRef={dragTabIdRef}
-            dragFolderIdRef={dragFolderIdRef}
-            isDragging={isDragging}
-            onBeforeReorder={onBeforeReorder}
-            lastSwapRef={lastSwapRef}
-            lastSwapTimeRef={lastSwapTimeRef}
-            lastFolderSwapRef={lastFolderSwapRef}
-            lastFolderSwapTimeRef={lastFolderSwapTimeRef}
-            disableEntryAnimation={disableEntryAnimation}
-            renamingFolderId={renamingFolderId}
-            depth={0}
-            onContextMenu={onContextMenu}
-          />
-        ),
-      )}
-    </>
-  );
-}
-
-function PinnedTabButton({
-  pt,
-  tab,
-  isActive,
-  onContextMenu,
-}: {
-  pt: { id: TabId; url: string; title: string; favicon: string };
-  tab: Tab | undefined;
-  isActive: boolean;
-  onContextMenu?: (items: ContextMenuItem[], e: React.MouseEvent) => void;
-}) {
-  const customTitle = useTabCustomizationStore((s) => s.customizations.get(pt.id))?.title;
-  const displayTitle = customTitle || pt.title || pt.url;
-  const handleContextMenu = (e: React.MouseEvent) => {
-    if (!onContextMenu) return;
-    const items: ContextMenuItem[] = [];
-    items.push(
-      {
-        label: "Unpin tab",
-        icon: "thumbtack-slash",
-        onSelect: () => sendFolderCommand(PINNED_TABS_TOGGLE_PIN, { tabId: pt.id }),
-      },
-      {
-        label: "Close tab",
-        icon: "xmark",
-        onSelect: () => sendCommand(TABS_CLOSE, { tabId: pt.id }),
-      },
-      {
-        label: "Customize tab",
-        icon: "sliders",
-        onSelect: () => sendCommand(TAB_CUSTOMIZATION_OPEN, { tabId: pt.id }),
-      },
-    );
-    if (tab?.url && tab.url !== pt.url) {
-      items.push({
-        label: "Restore original URL",
-        icon: "arrow-rotate-left",
-        onSelect: () => sendCommand(TABS_NAVIGATE, { tabId: pt.id, url: pt.url }),
-      });
-    }
-    onContextMenu(items, e);
-  };
-
-  return (
-    <button
-      type="button"
-      className={`flex flex-1 items-center justify-center cursor-pointer transition-colors duration-150 ${isActive ? "bg-glass-active" : "bg-glass-subtle hover:bg-glass-hover active:bg-glass-pressed"}`}
-      style={{
-        height: 32,
-        minWidth: 0,
-        borderRadius: "var(--radius-md)",
-        border: "none",
-        boxShadow: isActive ? "var(--shadow-subtle)" : undefined,
-      }}
-      tabIndex={-1}
-      onClick={() => sendCommand(PINNED_TABS_ACTIVATE, { tabId: pt.id })}
-      onContextMenu={handleContextMenu}
-      data-pinned-tab={pt.id}
-      data-tip={displayTitle}
-      aria-label={displayTitle}
-    >
-      <Favicon tab={tab ?? pt} />
-    </button>
-  );
-}
-
-export function PinnedTabsStrip({
-  pinnedTabs,
-  tabs,
-  activeTabId,
-  onContextMenu,
-}: {
-  pinnedTabs: { id: TabId; url: string; title: string; favicon: string }[];
-  tabs: Map<TabId, Tab>;
-  activeTabId: TabId | null;
-  onContextMenu?: (items: ContextMenuItem[], e: React.MouseEvent) => void;
-}) {
-  return (
-    <div
-      className="flex"
-      style={{
-        gap: "0.5rem",
-        padding: "0 0.375rem 0.25rem",
-      }}
-    >
-      {pinnedTabs.map((pt) => (
-        <PinnedTabButton
-          key={pt.id}
-          pt={pt}
-          tab={tabs.get(pt.id)}
-          isActive={pt.id === activeTabId}
-          onContextMenu={onContextMenu}
-        />
-      ))}
-    </div>
-  );
-}
-
-function SectionDropZone({
-  targetBookmarked,
-  dragTabIdRef,
-  dragFolderIdRef,
-  visible,
-  onBeforeReorder,
-  targetFolderId,
-}: {
-  targetBookmarked: boolean;
-  dragTabIdRef: React.RefObject<TabId | null>;
-  dragFolderIdRef?: React.RefObject<FolderId | null>;
-  visible: boolean;
-  onBeforeReorder?: () => void;
-  targetFolderId?: FolderId | null;
-}) {
-  const [over, setOver] = useState(false);
-  return (
-    <div
-      style={{
-        height: visible ? (over ? 32 : 24) : 0,
-        margin: visible ? "0.25rem 0.375rem" : "0 0.375rem",
-        borderRadius: "var(--radius-md)",
-        background: over
-          ? "oklch(var(--accent-L) var(--accent-C) var(--accent-hue, 250) / 0.1)"
-          : "transparent",
-        border: over
-          ? "1px solid oklch(var(--accent-L) var(--accent-C) var(--accent-hue, 250) / 0.3)"
-          : "1px solid transparent",
-        overflow: "hidden",
-        transition:
-          "height var(--duration-normal) var(--ease-in-out), margin var(--duration-normal) var(--ease-in-out), background var(--duration-fast) var(--ease-in-out), border-color var(--duration-fast) var(--ease-in-out)",
-      }}
-      onDragOver={(e) => {
-        e.preventDefault();
-        e.dataTransfer.dropEffect = "move";
-        setOver(true);
-      }}
-      onDragLeave={() => setOver(false)}
-      onDrop={(e) => {
-        e.preventDefault();
-        setOver(false);
-        // Handle folder drop — move to root level
-        if (dragFolderIdRef?.current) {
-          const folderId = dragFolderIdRef.current;
-          dragFolderIdRef.current = null;
-          onBeforeReorder?.();
-          sendFolderCommand(FOLDERS_REORDER, {
-            folderId,
-            parentFolderId: null,
-          });
-          return;
-        }
-        // Handle tab drop
-        const tabId = dragTabIdRef.current;
-        dragTabIdRef.current = null;
-        if (!tabId) return;
-        onBeforeReorder?.();
-        sendCommand(TABS_REORDER, {
-          tabId,
-          targetBookmarked,
-          targetFolderId: targetFolderId ?? null,
-        });
-      }}
-    />
-  );
-}
-
-function TabSection({
-  bookmarked,
-  bookmarkedTree,
-  ephemeral,
-  activeTabId,
-  exitingIds,
-  dragTabIdRef,
-  dragFolderIdRef,
-  isDragging,
-  onClearEphemeral,
-  disableEntryAnimation,
-  renamingFolderId,
-  onContextMenu,
-}: {
-  bookmarked: Tab[];
-  bookmarkedTree: TreeItem[];
-  ephemeral: Tab[];
-  activeTabId: TabId | null;
-  exitingIds: Set<TabId>;
-  dragTabIdRef: React.RefObject<TabId | null>;
-  dragFolderIdRef: React.RefObject<FolderId | null>;
-  isDragging: boolean;
-  onClearEphemeral: () => void;
-  disableEntryAnimation?: boolean;
-  renamingFolderId: FolderId | null;
-  onContextMenu?: (items: ContextMenuItem[], e: React.MouseEvent) => void;
-}) {
-  // Divider visible when ephemeral tabs exist or during drag; lingers for fade-out
-  const dividerTarget = ephemeral.length > 0 || isDragging;
-  const [showDivider, setShowDivider] = useState(dividerTarget);
-  useEffect(() => {
-    if (dividerTarget) {
-      setShowDivider(true);
-      return;
-    }
-    const t = setTimeout(() => setShowDivider(false), 200); // match --duration-normal
-    return () => clearTimeout(t);
-  }, [dividerTarget]);
-
-  // ── FLIP animation for smooth tab reordering ──
-  const flipContainerRef = useRef<HTMLDivElement>(null);
-  const prevRectsRef = useRef<Map<string, DOMRect>>(new Map());
-
-  const snapshotPositions = useCallback(() => {
-    const container = flipContainerRef.current;
-    if (!container) return;
-    const rects = new Map<string, DOMRect>();
-    for (const el of container.querySelectorAll<HTMLElement>("[data-tab-id]")) {
-      const id = el.dataset.tabId;
-      if (id) rects.set(id, el.getBoundingClientRect());
-    }
-    prevRectsRef.current = rects;
-  }, []);
-
-  useLayoutEffect(() => {
-    const container = flipContainerRef.current;
-    const prev = prevRectsRef.current;
-    if (!container || prev.size === 0) return;
-    prevRectsRef.current = new Map();
-
-    const elements = container.querySelectorAll<HTMLElement>("[data-tab-id]");
-    const toAnimate: { el: HTMLElement; deltaY: number }[] = [];
-
-    for (const el of elements) {
-      const id = el.dataset.tabId;
-      if (!id) continue;
-      const oldRect = prev.get(id);
-      if (!oldRect) continue;
-      const newRect = el.getBoundingClientRect();
-      const deltaY = oldRect.top - newRect.top;
-      if (Math.abs(deltaY) < 1) continue;
-      toAnimate.push({ el, deltaY });
-    }
-
-    if (toAnimate.length === 0) return;
-
-    for (const { el, deltaY } of toAnimate) {
-      el.style.transform = `translateY(${deltaY}px)`;
-      el.style.transition = "none";
-    }
-
-    void container.offsetHeight;
-    requestAnimationFrame(() => {
-      for (const { el } of toAnimate) {
-        el.style.transition = "transform 200ms cubic-bezier(0.25, 0.1, 0.25, 1)";
-        el.style.transform = "";
-      }
-      setTimeout(() => {
-        for (const { el } of toAnimate) {
-          el.style.removeProperty("transition");
-          el.style.removeProperty("transform");
-        }
-      }, 210);
-    });
-  });
-
-  // ── Drag swap tracking ──
-  const lastSwapRef = useRef<{ targetId: TabId; position: string } | null>(null);
-  const lastSwapTimeRef = useRef(0);
-  const lastFolderSwapRef = useRef<{ targetId: FolderId; position: string } | null>(null);
-  const lastFolderSwapTimeRef = useRef(0);
-
-  return (
-    <div ref={flipContainerRef}>
-      {bookmarked.length > 0 || bookmarkedTree.some((item) => item.type === "folder") ? (
-        <>
-          <BookmarkedTree
-            tree={bookmarkedTree}
-            activeTabId={activeTabId}
-            exitingIds={exitingIds}
-            dragTabIdRef={dragTabIdRef}
-            dragFolderIdRef={dragFolderIdRef}
-            isDragging={isDragging}
-            onBeforeReorder={snapshotPositions}
-            lastSwapRef={lastSwapRef}
-            lastSwapTimeRef={lastSwapTimeRef}
-            lastFolderSwapRef={lastFolderSwapRef}
-            lastFolderSwapTimeRef={lastFolderSwapTimeRef}
-            disableEntryAnimation={disableEntryAnimation}
-            renamingFolderId={renamingFolderId}
-            onContextMenu={onContextMenu}
-          />
-          {/* Root-level drop zone — visible during drag so tabs/folders can be placed outside folders */}
-          <SectionDropZone
-            targetBookmarked={true}
-            dragTabIdRef={dragTabIdRef}
-            dragFolderIdRef={dragFolderIdRef}
-            visible={isDragging}
-            onBeforeReorder={snapshotPositions}
-          />
-        </>
-      ) : (
-        <SectionDropZone
-          targetBookmarked={true}
-          dragTabIdRef={dragTabIdRef}
-          dragFolderIdRef={dragFolderIdRef}
-          visible={isDragging}
-          onBeforeReorder={snapshotPositions}
-        />
-      )}
-
-      {/* Ephemeral divider — fades in/out based on content or drag state */}
-      <div
-        className="flex items-center"
-        style={{
-          gap: "0.5rem",
-          padding: showDivider ? "0.375rem 0.5rem 0.25rem 0.875rem" : "0 0.5rem 0 0.875rem",
-          margin: "0.125rem 0",
-          height: showDivider ? undefined : 0,
-          opacity: showDivider ? 1 : 0,
-          overflow: "hidden",
-          transition:
-            "opacity var(--duration-normal) var(--ease-in-out), height var(--duration-normal) var(--ease-in-out), padding var(--duration-normal) var(--ease-in-out)",
-        }}
-      >
-        <div style={{ flex: 1, height: 1, background: "var(--glass-border)" }} />
-        {ephemeral.length > 0 && (
-          <button
-            type="button"
-            className="flex items-center cursor-pointer bg-transparent text-glass-text-hint transition-colors duration-150 hover:bg-glass-hover hover:text-glass-text-hover"
-            style={{
-              gap: "0.25rem",
-              border: "none",
-              fontSize: "var(--text-xs)",
-              fontWeight: 500,
-              fontFamily: "inherit",
-              padding: "0.125rem 0.375rem",
-              borderRadius: "var(--radius-sm)",
-              whiteSpace: "nowrap",
-              letterSpacing: "0.02em",
-            }}
-            tabIndex={-1}
-            onClick={onClearEphemeral}
-            data-tip="Clear ephemeral tabs"
-          >
-            Clear <Icon name="broom" />
-          </button>
-        )}
-      </div>
-
-      {ephemeral.length > 0 ? (
-        ephemeral.map((tab) => (
-          <TabItem
-            key={tab.id}
-            tab={tab}
-            isActive={tab.id === activeTabId}
-            isEphemeral={true}
-            exiting={exitingIds.has(tab.id)}
-            isBookmarkedSection={false}
-            dragTabIdRef={dragTabIdRef}
-            isDragged={tab.id === (isDragging ? dragTabIdRef.current : null)}
-            isDragging={isDragging}
-            onBeforeReorder={snapshotPositions}
-            lastSwapRef={lastSwapRef}
-            lastSwapTimeRef={lastSwapTimeRef}
-            disableEntryAnimation={disableEntryAnimation}
-            onContextMenu={onContextMenu}
-          />
-        ))
-      ) : (
-        <SectionDropZone
-          targetBookmarked={false}
-          dragTabIdRef={dragTabIdRef}
-          visible={isDragging}
-          onBeforeReorder={snapshotPositions}
-        />
-      )}
-    </div>
-  );
-}
-
-// ── Main Component ──────────────────────────────────────────────
+// ── Resize hook ─────────────────────────────────────────────────
 
 function useSidebarResize(sidebarWidth: number) {
   const [dragWidth, setDragWidth] = useState<number | null>(null);
@@ -1524,6 +195,8 @@ function useSidebarResize(sidebarWidth: number) {
   };
 }
 
+// ── Main Component ──────────────────────────────────────────────
+
 export function SidebarPanel() {
   const visible = useSidebarStore((s) => s.visible);
   const announcement = useSidebarStore((s) => s.announcement);
@@ -1531,7 +204,7 @@ export function SidebarPanel() {
   const tabs = useTabsStore((s) => s.tabs);
   const activeTabId = useTabsStore((s) => s.activeTabId);
   const pinnedTabs = usePinnedTabsStore((s) => s.pinnedTabs);
-  const pinnedTabIds = new Set(pinnedTabs.map((p) => p.id));
+  const pinnedTabIds = useMemo(() => new Set(pinnedTabs.map((p) => p.id)), [pinnedTabs]);
   const workspaces = useWorkspacesStore((s) => s.workspaces);
   const activeWorkspaceId = useWorkspacesStore((s) => s.activeWorkspaceId);
   const folders = useFoldersStore((s) => s.folders);
@@ -1612,7 +285,7 @@ export function SidebarPanel() {
 
   const handleClearEphemeral = () => {
     if (activeWorkspaceId) {
-      sendCommand(TABS_CLEAR_EPHEMERAL, { workspaceId: activeWorkspaceId });
+      window.chiaroscuro.sendCommand(TABS_CLEAR_EPHEMERAL, { workspaceId: activeWorkspaceId });
     }
   };
