+++ src/features/sidebar/sidebar.store.ts
@@ -1,7 +1,8 @@
 import { create } from "zustand";
-import { TABS_LIST_CHANGED, type TabsListChangedEvent } from "../tabs/tabs.shared";
-import { WORKSPACES_SWITCHED, type WorkspacesSwitchedEvent } from "../workspaces/workspaces.shared";
-import { SIDEBAR_VISIBILITY_CHANGED, type SidebarVisibilityChangedPayload } from "./sidebar.shared";
+import { typedOnEvent } from "../../shared/typed-on-event";
+import { TABS_LIST_CHANGED, type TabsEvents } from "../tabs/tabs.shared";
+import { WORKSPACES_SWITCHED, type WorkspacesEvents } from "../workspaces/workspaces.shared";
+import { SIDEBAR_VISIBILITY_CHANGED, type SidebarEvents } from "./sidebar.shared";
 
 interface SidebarState {
   visible: boolean;
@@ -18,18 +19,17 @@ let lastTabCount = 0;
 export function subscribeToEvents(
   onEvent: (name: string, callback: (payload: unknown) => void) => () => void,
 ): () => void {
+  const on = typedOnEvent<SidebarEvents & WorkspacesEvents & TabsEvents>(onEvent);
   const unsubs: (() => void)[] = [];
 
   unsubs.push(
-    onEvent(SIDEBAR_VISIBILITY_CHANGED, (payload) => {
-      const { visible } = payload as SidebarVisibilityChangedPayload;
+    on(SIDEBAR_VISIBILITY_CHANGED, ({ visible }) => {
       useSidebarStore.setState({ visible });
     }),
   );
 
   unsubs.push(
-    onEvent(WORKSPACES_SWITCHED, (payload) => {
-      const { workspaceName } = payload as WorkspacesSwitchedEvent;
+    on(WORKSPACES_SWITCHED, ({ workspaceName }) => {
       const count = lastTabCount;
       useSidebarStore.setState({
         announcement: `Switched to ${workspaceName}, ${count} tab${count !== 1 ? "s" : ""}`,
@@ -38,8 +38,7 @@ export function subscribeToEvents(
   );
 
   unsubs.push(
-    onEvent(TABS_LIST_CHANGED, (payload) => {
-      const { tabs } = payload as TabsListChangedEvent;
+    on(TABS_LIST_CHANGED, ({ tabs }) => {
       const count = tabs.length;
       if (lastTabCount > 0 && lastTabCount !== count) {
         useSidebarStore.setState({
